move lexer tokens to separate file
